+++ src/main/java/com/interview/notes/code/year/y2024/jan24/test14/DiscountedPriceFetcher3.java
@@ -46,7 +46,7 @@ public class DiscountedPriceFetcher3 {
                 int discount = extractValue(data, "discount");
 
                 int discountedPrice = price - (discount * price / 100);
-                return (int) Math.round(discountedPrice);
+                return Math.round(discountedPrice);
             } else {
                 return -1;
             }
